Move core classes into core direcotry and fix all warnings

+++ midori_engine/src/midori/renderer/Buffer.h
@@ -46,7 +46,7 @@ namespace midori {
         std::string Name;
         ShaderDataType Type;
         uint32_t Size;
-        uint32_t Offset;
+        uint64_t Offset;
         bool Normalized;
 
         BufferElement() = default;
